feat: IncomeRequest 구현 및 공통 request 필드 분리
