+++ src/main/java/com/poortorich/income/request/IncomeRequest.java
@@ -1,0 +1,9 @@
+package com.poortorich.income.request;
+
+import com.poortorich.accountbook.request.AccountBookRequest;
+import lombok.experimental.SuperBuilder;
+
+@SuperBuilder
+public class IncomeRequest extends AccountBookRequest {
+
+}
